Move first integration test to the new latex view test module
